rename main and run to cli for consistency
